requirements.txt
alembic==1.14.1
blinker==1.8.2
click==8.1.8
Flask==3.0.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==2.1.5
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.0.6
WTForms==3.1.2
requests==2.31.0
authlib==1.3.0
pandas==2.0.3
numpy==1.24.4
python-dotenv==1.0.1
email-validator==2.2.0
pymysql==1.0.2
gunicorn
Flask-Session